Make Int arbitrary precision, as the spec always said it was
docs/language-spec.md states twice that Int is arbitrary precision and that
users should not have to think about overflow. The implementation was i64, and
a benchmark campaign found the gap: across 142 Exercism tasks Aven refused 6
cases that Python and Ruby both express, four of them values beyond 64-bit.

num-bigint, with no small-integer fast path. The spec calls the runtime
representation an implementation detail and performance is deferred to the VM,
so a tagged fast path now would be complexity in exactly the place the core
principle warns about.

Two standing bugs disappear as a consequence rather than being fixed: i64::MIN
finally has a literal spelling, since unary `-` no longer applies to a positive
literal that has already overflowed; and `check` and `run` agree, because the
range check they disagreed about no longer exists.

At the boundaries, both directions are now lossless or loud. JSON integers
beyond i64 round-trip exactly instead of demoting to @Float — `2^64-1` used to
come back as 1.8446744073709552e19. A host function that wants a Rust i64 and
is handed something larger gets a typed error naming the value, never a silent
truncation.

Verified by hand rather than trusted: the run was killed before it could claim
its gates, so I ran all three myself — build clean, 29 test suites green,
clippy silent under -D warnings — and checked the acceptance criteria against
the binary, including 2^100 and a live JSON round-trip.

+++ crates/aven-check/src/comptime.rs
@@ -2,7 +2,7 @@ use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
 use std::sync::{Mutex, OnceLock};
 
-use aven_core::{Diagnostic, Label, Span, codes};
+use aven_core::{Diagnostic, Int, Label, Span, codes};
 use aven_parser::{Expr, ExprKind, Literal, MatchArm, Param, decode_string_literal};
 
 use crate::checker::string_literal_label;
@@ -171,7 +171,7 @@ impl From<RowTail> for CanonicalRowTail {
 #[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
 enum CanonicalLiteral {
     Bool(bool),
-    Int(i64),
+    Int(Int),
     Float(u64),
     InvalidNumber(String),
     String(String),
@@ -198,7 +198,7 @@ fn canonical_number(text: &str) -> CanonicalLiteral {
         );
     }
 
-    normalized.parse::<i64>().map_or_else(
+    normalized.parse::<Int>().map_or_else(
         |_| CanonicalLiteral::InvalidNumber(text.to_owned()),
         CanonicalLiteral::Int,
     )
